app/backend/requirements.txt
alembic==1.16.4
black==23.7.0
blinker==1.9.0
click==8.2.1
flake8==6.1.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.0.5
greenlet==3.2.3
gunicorn==23.0.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
mysqlclient==2.2.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.10.1
pytest==7.4.0
python-dotenv==1.0.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
Werkzeug==3.1.3
Pillow>=9.0.0
pymysql
psycopg2-binary




